I tried doing view frustrum culling but its harder than I thought.
I can't do it the "simple" block way because if I point the camera downwards the camera x, z "towards" are 0, 0.
The temporary fix is to always draw things within a certain radius around youself - this gets broken as you climb up high places, put your ass in the air and look down.

The clever way would be to use xfov and yfov, but you tell me how to do that...

going to bed.